Fix build errors and some formatting
