+++ src/GitHubVulnerabilities2Db/Gallery/ThrowingTelemetryService.cs
@@ -413,5 +413,10 @@ namespace GitHubVulnerabilities2Db.Gallery
         {
             throw new NotImplementedException();
         }
+
+        public void TrackAdminApiReflow(int packageCount, int acceptedCount, string reason, string callerAppId)
+        {
+            throw new NotImplementedException();
+        }
     }
 }
